attempt to fix a problem with disappearing node labels
Occasionally the labels generated by this plugin will disappear and the node
has to be disconnected and reconnected in order to bring them back.  My current
working theory is that the WeakHashMap being used as a label cache is being
cleared of values when we aren't expecting it to be and labels are not
re-generated on demand when this happens.

This commit does not remedy the fact that the unit tests no longer work.

Also updated the version of Jenkins that the plugin is built against to
something not quite so ancient.