Only call uname-m if java os.arch reports x86
If java os.arch is 'x86', there is a reasonable chance that the agent
is running a 32 bit JVM on a 64 bit (Windows) machine.  Avoid the call to
'uname -m' in other cases.

Yes, it is possible someone found a way to run a 32 bit JVM on a 64 bit
ARM configuration.  No, this does not support such a configuration.