[JENKINS-52798] Support Alpine Linux
Alpine Linux does not support the lsb_release command. The lsb_release
command is used by the platformlabeler-plugin to determine the name
of the Linux distributor. For example, it is used to determine CentOS,
Debian, OpenSUSE, Red Hat, SUSE, and Ubuntu as Linux operating system
distributions.

The platformlabeler plugin should support Alpine Linux since it is the
basis of one of the three Jenkins Docker images and is increasingly the
preferred image due to its small size.