[JENKINS-17904] allow to configure labels per node
in certain cases one doesn't want to have all the generated labels one
certain nodes.
Additionally add a global configuration for the labels

This might also fix JENKINS-17615.
The problem might be that the Node object gets detroyed and recreated
when a node config changes while the associated Computer is untouched
when the connection itself is unchanged (e.g. just a label change). I
noticed that when chaning the nodeproperty my changes didn't get
reflected. Only after caching the computer it started working.