Use SPECIAL_CASE_ARCH to identify x86 special case tests
The special case tests are run on Intel processors to validate x86
detection on computers where the processor may be 64 bit but the Java
process is running as a 32 bit process.

Hypothetically the same condition is possible on some ARM processors,
but ARM processors are almost always running a Java version that matches
the 32 / 64 bitness of the operating system.