Disable Dependabot eval of Jenkins core and plugins
It is better for users if updates of Jenkins core dependency and
Jenkins plugin dependencies are considered individually and based on
more detailed analysis.

A new release of Jenkins is almost always compatible with the release on
which the plugin depends. A new Jenkins release should become required
when there are significant changes in the Jenkins core release that
are needed by the plugin.  For example, Jenkins 2.164.1 was the first
long term support release to support Java 11. The decision to upgrade
the plugin to require Jenkins 2.164.1 was a decision to include Java 11
support fully.

The same logic applies to Jenkins plugin releases. When the
platformlabeler plugin depends on a specific plugin version, it mandates
that the user must install at least that version.  When the version
of the plugin dependency is incremented without a compelling reason,
it limits the choice of plugin users to that plugin version or newer.