[JENKINS-61597] Use freebsd-version for jail version number
Using the uname output on FreeBSD provides the kernel version.  The
userland version is much more relevant, especially in FreeBSD jails.
This changes the behavior to label the agent with the userland
version rather than the kernel version.

This change makes FreeBSD jail behaviors closer to the behavior of
Docker images.  When a Docker image reports its operating system,
it reads the userland values, not the kernel values.