Assign locale in lower case conversions
Strings are expected to be English language in general and should not risk
applying the case conversion of the default locale.  The default locale
case conversion would be a problem if it were Turkish and the string LINUX
were converted to l<i>nux where the <i> is a letter i without the dot.