Detect debian_version data file correctly
Previously was looking at the basename of the data file argument
(releaseFile).  In the Debian case, the basename of the data file is
'os-release'.  The replacement data files in the Red Hat and SuSE
cases have different names.

The Debian case uses an additional file that is consulted after
reading the original data from the os-release file.  The Debian case
does not replace the os-release file, it augments it with more
information.