Include Debian unstable and testing in os-release docs
When the Debian unstable and Debian testing releases are installed
without the lsb-core package, the lsb_release command is not available.
When the lsb_release command is not available, the automatic labeling
uses /etc/os-release and /etc/debian_version instead.  The contents
of /etc/debian_version are different than the values reported by
`lsb_release -r`.  Document those differences.