Enable OpenClover coverage reporting
OpenClover provides coverage reporting that is similar to other
coverage reporting tools, like JaCoCo.  In addition, it provides a test
optimization facility that may avoid unmodified tests when evaluating
a pull request or other change.

Tests in this plugin are fast enough that the test avoidance does not
matter.  Tests in other plugins might benefit significantly from test
avoidance, if it is accurate enough.