Move forkCount to a CI profile
James Nord noted that setting the surefire plugin forkCount in the
configuration of all builds may be a disadvantage for developers with
multi-core laptops that they do not want to heavily load with tests
when running a plugin build.

He suggested a profile for continuous integration environments that is
not enabled for the typical developer experience.

This is exploring the idea of a profile that is specific to continuous
integration.