Add os-release test data for Amazon Linux 2022
Amazon Linux 2022 enables selinux by default.  That may cause problems
for the git client plugin and for other security related operations.

The generator script for the os-release test data generation needs to
be reworked to do a better job of extracting the os-release file from
the correct image.  It does not extract Amazon Linux 2022 data currently
and it still has conditionals for Scientific Linux 6.