Require Jenkins 2.332.4 or newer
Security advisory has been published for Jenkins 2.332.3 and
older.  No benefit for users to support releases with known
security issues.  If they need an older release of Jenkins,
they can also choose to use an older release of the plugin.

Increasing the minimum Jenkins version also more accurately
describes the range of versions that are tested.