[JENKINS-69814] Assign Debian unstable a valid release
Debian unstable modified the lsb_release output in the Oct 2022
Docker container image to report "n/a" as the release value.
That is a sensible and reasonable value, but it collides with the
desire of the platform labeler plugin to not assign labels that
contain "/" and it breaks with what people might expect for
compatibility with previous behavior.

Set the release value to "unstable" if Debian and release is "n/a".