[JENKINS-69971] Assign Debian testing label correctly
Debian testing has periods during its lifecycle when the lsb_release
command cannot distinguish between testing and unstable.  That is an
accepted condition for the Debian developers.

See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=845651

When in that state, the apt-cache command can be used to determine the
Debian release that is running.

The base-files package is used to query the apt-cache policy because it
is required for all Debian installations and because it is the package
that provides the /etc/os-release file that is used elsewhere to determine
version information in case the lsb_release command is not available.