[JENKINS-70737] Detect Debian unstable even during freeze
The platform labeler incorrectly reports Debian unstable ("sid") as
Debian 12 instead of reporting it as Debian unstable.  Debian 12 is
still in one of its freeze.

The Debian version guesser that had been implemented previously was
only being used in those cases where `lsb_release -a` returns "n/a"
as the version.  That works when the unstable line is not preparing
for a release.  When the unstable line is preparing for a release,
the contents of `/etc/os-release` and the output of `lsb_release -a`
cannot be distinguished from the final release version.

Call the `apt-cache policy` command always to get the distribution that
provides the policy for the base operating system package.