[JENKINS-70737] Report Debian unstable correctly during freeze
During soft freeze and hard freeze before a new Debian release,
the /etc/os-release file on Debian unstable intentionally
looks like the next major release.  That is intentional.  See
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=845651 for details.

When in that state, the apt-cache command can be used to determine the
Debian release that is running.

The base-files package is used to query the apt-cache policy because it
is required for all Debian installations and because it is the package
that provides the /etc/os-release file that is used elsewhere to determine
version information in case the lsb_release command is not available.

This changes to call `apt-cache policy base-files` on all Debian machines.
If the return value is not `unstable` or `testing` or `sid` then it
is ignored.

Adapt PlatformDetailsTaskTest to unstable in freeze