Update Debian testing & unstable test data
Update sample data for Debian GNU/Linux 12 (bookworm) in the testing
distribution and in the unstable distribution.

Condition is temporary until the release of Debian 12 (bookworm).
Once Debian 12 is released, then the release will return to "n/a" and
the test code will need to be adapted to the change.