Remove RHEL 7 and derivatives from docs and tests
https://www.jenkins.io/blog/2023/05/30/operating-system-end-of-life/
describes some of the reasons to stop documenting and to stop testing
with Red Hat Enterprise Linux 7.  The container images used for the
operating system test data have not been updated in over 2 years.
The operating system will be unsupported by the Jenkins project beginning
in November 2023.  Removing it from documentation does not break the
software that supports it, but it stops wasting time testing it and
tracking its progress.