Create a Windows Server label with os.name (#1201)
The Jenkins agent list shows the value of the Java `os.name` property to
users in the "Architecture" column.  This change adds a new label for
agents based on the value in that architecture column.  It is enabled by
default.

For most Linux distributions (CentOS, Debian, openSUSE, Red Hat,
Ubuntu), the value will be "Linux".  FreeBSD agents will have the value
"FreeBSD".  Windows agents will have a value that depends on the Windows
version that is installed.  Some examples include:

* Windows11
* Windows10
* WindowsServer2016
* WindowsServer2019
* WindowsServer2023

The operating system name label can be enabled and disabled globally and
can be enabled and disabled on each agent individually.

Fixes #612

Existing labels are unaffected by the addition.  Windows agents will
still be labeled "windows", as controlled by the "Generate label with
OS name" setting.  Debian agents will still be labeled "Debian".

Interactive testing has been performed to verify Windows and Linux
agents are labeled as expected with the various configurations.