Drop support for Amazon Linux 2
Amazon Linux 2 is a derivative of Red Hat Enterprise Linux 7 and uses
the OpenSSH version that is provided with Red Hat Enterprise Linux 7.
https://www.jenkins.io/blog/2023/05/30/operating-system-end-of-life/
announced the Nov 16, 2023 end of life for Red Hat Enterprise Linux 7
and all its derivatives.

The change does not alter any production code but removes the test data
and maintenance content that checked Amazon Linux 2.